fix: revert back to using the app dir
