+++ apps/website/next.config.js
@@ -2,6 +2,7 @@
 module.exports = {
 	reactStrictMode: true,
 	swcMinify: true,
+	experimental: { appDir: true },
 	async redirects() {
 		return [
 			{
