--- apps/website/src/pages/_document.tsx
@@ -1,13 +1,0 @@
-import { Html, Main, Head, NextScript } from 'next/document';
-
-export default function Document() {
-	return (
-		<Html lang="en">
-			<Head />
-			<body>
-				<Main />
-				<NextScript />
-			</body>
-		</Html>
-	);
-}
